add Twitter component & fix file structure
