+++ src/components/atoms/Twitter/index.ts
@@ -1,0 +1,1 @@
+export { Twitter } from "./Twitter";
