+++ src/components/organisms/Base.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import { Footer } from "../organisms/Footer";
+import { Header } from "@Components/molecules/Header";
+import { Footer } from "@Components/molecules/Footer";
 
 type PropsType = {
   children: React.ReactNode;
@@ -8,6 +9,7 @@ type PropsType = {
 const Base: React.FC<PropsType> = (props) => {
   return (
     <div className="min-h-screen m-12">
+      <Header />
       <div className="container mx-auto my-12">{props.children}</div>
       <Footer />
     </div>
