+++ src/components/organisms/PostBase.tsx
@@ -1,6 +1,5 @@
 import React from "react";
-import { Header } from "../organisms/Header";
-import { Footer } from "../organisms/Footer";
+import { Footer } from "@Components/molecules/Footer";
 
 type PropsType = {
   children: React.ReactNode;
@@ -9,7 +8,6 @@ type PropsType = {
 const Base: React.FC<PropsType> = (props) => {
   return (
     <div className="min-h-screen m-12">
-      <Header />
       <div className="container mx-auto my-12">{props.children}</div>
       <Footer />
     </div>
